r600g: avoid unnecessary shader exports
In some cases TGSI shader has more color outputs than the number of CBs,
so it seems we need to limit the number of color exports. This requires
different shader variants depending on the nr_cbufs, but on the other hand
we are doing less exports, which are very costly.

Signed-off-by: Vadim Girlin <vadimgirlin@gmail.com>